# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Nervum API
  description: |
    API for Nervum – environment maps SaaS (visualize and manage services, databases,
    queues, and relationships). This backend powers the Nervum web app and can be
    used for integrations.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/organizations:
    post:
      tags: [Organizations]
      summary: Create organization
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Organizations]
      summary: List organizations
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/organizations/{id}:
    get:
      tags: [Organizations]
      summary: Get organization by ID
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Organizations]
      summary: Update organization
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpdate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Organizations]
      summary: Delete organization
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/invitations:
    post:
      tags: [Invitations]
      summary: Create invitation
      security: [{ cookieAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Invitations]
      summary: List invitations
      security: [{ cookieAuth: [] }]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, expired]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invitation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/invitations/{id}:
    delete:
      tags: [Invitations]
      summary: Delete invitation
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/UUIDPathId'
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/invitations/by-token/{token}:
    get:
      tags: [Invitations]
      summary: Get invitation by token (public)
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationByTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Invitation not found or expired
        '410':
          description: Invitation already used or expired
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/invitations/accept:
    post:
      tags: [Invitations]
      summary: Accept invitation (public)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationAcceptRequest'
      responses:
        '200':
          description: Existing user; session started
          headers:
            Set-Cookie:
              description: Session cookie (nervum_session)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '201':
          description: New user created; session started
          headers:
            Set-Cookie:
              description: Session cookie (nervum_session)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Invitation not found or expired
        '409':
          description: Invitation already used
        '410':
          description: Invitation expired
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: nervum_session
      description: Session ID (UUID) set by login/register/accept-invitation
  parameters:
    UUIDPathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request (e.g. validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [admin, manager, member]
        organization_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        website:
          type: string
        owner_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrganizationCreate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        website:
          type: string
    OrganizationUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        website:
          type: string
...
    Invitation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        token:
          type: string
        email:
          type: string
        organization_id:
          type: string
          format: uuid
        invited_by_id:
          type: string
          format: uuid
        role:
          type: string
        environment_id:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, accepted, expired]
        accepted_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        teams:
          type: array
          items:
            type: object
    InvitationCreate:
      type: object
      required: [email, team_ids]
      properties:
        email:
          type: string
          format: email
        team_ids:
          type: array
          items:
            type: string
            format: uuid
        environment_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [admin, manager, member]
    InvitationCreateResponse:
      type: object
      properties:
        invitation:
          $ref: '#/components/schemas/Invitation'
        invite_url:
          type: string
        token:
          type: string
    InvitationByTokenResponse:
      type: object
      properties:
        email:
          type: string
        organization_id:
          type: string
        organization_name:
          type: string
        team_ids:
          type: array
          items:
            type: string
        environment_id:
          type: string
        role:
          type: string
        expires_at:
          type: string
          format: date-time
    InvitationAcceptRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
        name:
          type: string
        password:
          type: string
          minLength: 8
          description: Required for new users; min 8 characters
...
security:
  - cookieAuth: []